add first tests for app
